feat: finale
