+++ src/app/core/utils/input-validators.ts
@@ -1,0 +1,6 @@
+export function blockInvalidKeys(event: KeyboardEvent) {
+  const invalidKeys = ['-', '+', 'e', 'E', '.', ',', '0'];
+  if (invalidKeys.includes(event.key)) {
+    event.preventDefault();
+  }
+}
