--- src/app/domain/models/.keep
@@ -1,1 +1,0 @@
-// Placeholder para manter o diretório no controle de versão
