+++ src/app/presentation/pages/event-projection/event-projection.component.html
@@ -1,7 +1,7 @@
 <div class="page-wrapper">
   <div class="header">
     <h2>Iniciar novas Entidades</h2>
-    <button class="close-btn" aria-label="Fechar">×</button>
+    <button (click)="navigateToSplash()" class="close-btn" aria-label="Fechar">×</button>
   </div>
 
   <app-start-entities />
@@ -9,7 +9,7 @@
   <app-events-chart [data]="chartData()" />
 
   <div class="actions">
-    <button class="secondary-btn">Fechar</button>
+    <button class="secondary-btn" (click)="navigateToSplash()">Fechar</button>
     <button class="primary-btn" (click)="recalcularProjecao()">Iniciar novas entidades</button>
   </div>
 </div>
