--- src/app/shared/material/.keep
@@ -1,1 +1,0 @@
-// Placeholder para manter o diretório no controle de versão
